[telemetry] Add logging and additional check to StalePyc scripts.
Not certain of the exact problem, but maybe more logging will help. Using print instead of logging, because we haven't set the logging level at import time.


BUG=272333
TEST=None.

Review URL: https://chromiumcodereview.appspot.com/22796016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@218444 0039d316-1c4b-4281-b951-d872f2087c98

+++ tools/perf/perf_tools/__init__.py
@@ -2,7 +2,9 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 # Copyright (c) 2012 The Chromium Authors. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
+
 """A library for Chrome GPU test code."""
+
 import os
 import sys
 
@@ -19,13 +21,15 @@ def _RemoveAllStalePycFiles():
       pyc_path = os.path.join(dirname, filename)
       py_path = os.path.join(dirname, root + '.py')
       if not os.path.exists(py_path):
+        print >> sys.stderr, 'Removing stale .pyc file:', pyc_path
         os.remove(pyc_path)
 
-    if not os.listdir(dirname):
+    if os.path.isdir(dirname) and not os.listdir(dirname):
+      print >> sys.stderr, 'Removing empty directory:', dirname
       os.removedirs(dirname)
 
 
-def Init():
+def _Init():
   telemetry_path = os.path.join(os.path.dirname(__file__),
                                 os.pardir, os.pardir, 'telemetry')
   absolute_telemetry_path = os.path.abspath(telemetry_path)
@@ -37,4 +41,4 @@ def Init():
 
 
 _RemoveAllStalePycFiles()
-Init()
+_Init()
