+++ tools/telemetry/telemetry/__init__.py
@@ -2,9 +2,9 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 # Copyright (c) 2013 The Chromium Authors. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
-"""
-A library for cross-platform browser tests.
-"""
+
+"""A library for cross-platform browser tests."""
+
 import inspect
 import os
 import sys
@@ -43,9 +43,11 @@ def _RemoveAllStalePycFiles():
       pyc_path = os.path.join(dirname, filename)
       py_path = os.path.join(dirname, root + '.py')
       if not os.path.exists(py_path):
+        print >> sys.stderr, 'Removing stale .pyc file:', pyc_path
         os.remove(pyc_path)
 
-    if not os.listdir(dirname):
+    if os.path.isdir(dirname) and not os.listdir(dirname):
+      print >> sys.stderr, 'Removing empty directory:', dirname
       os.removedirs(dirname)
